+++ libs/designsystem/router-outlet/ng-package.json
@@ -1,0 +1,1 @@
+{}
