+++ libs/designsystem/router-outlet/src/router-outlet.component.scss
