+++ libs/designsystem/src/lib/index.ts
@@ -28,6 +28,7 @@ export * from '@kirbydesign/designsystem/item-sliding';
 export * from '@kirbydesign/designsystem/form-field';
 export * from '@kirbydesign/designsystem/avatar';
 export * from '@kirbydesign/designsystem/progress-circle';
+export * from '@kirbydesign/designsystem/router-outlet';
 
 export * from './scss/scss-helper';
