+++ libs/designsystem/tsconfig.spec.json
@@ -34,7 +34,8 @@
       "@kirbydesign/designsystem/list": ["list/index.ts"],
       "@kirbydesign/designsystem/form-field": ["form-field/index.ts"],
       "@kirbydesign/designsystem/avatar": ["avatar/index.ts"],
-      "@kirbydesign/designsystem/progress-circle": ["progress-circle/index.ts"]
+      "@kirbydesign/designsystem/progress-circle": ["progress-circle/index.ts"],
+      "@kirbydesign/designsystem/router-outlet": ["router-outlet/index.ts"]
     }
   },
   "files": ["test.ts"],
